refactor: move context menu to its own file
