+++ web/src/common/components/Menu/Menu.tsx
@@ -3,11 +3,7 @@
 import { MenuItem, type MenuItemProps, Menu as MuiMenu } from "@mui/material";
 import { Children, isValidElement } from "react";
 
-import { deleteNode } from "../../../modules/topic/store/actions";
-import { closeContextMenu } from "../../store/contextMenuActions";
-import { useAnchorPosition, useContextMenuContext } from "../../store/contextMenuStore";
-
-const addCloseOnClick = (closeMenu: () => void, children: React.ReactNode) => {
+export const addCloseOnClick = (closeMenu: () => void, children: React.ReactNode) => {
   return Children.map(children, (child) => {
     if (!isValidElement(child) || child.type !== MenuItem) {
       throw new Error("Menu children must be MenuItems");
@@ -41,39 +37,6 @@ export const Menu = ({ anchorEl, isOpen, closeMenu, children }: MenuProps) => {
 
   return (
     <MuiMenu anchorEl={anchorEl} open={isOpen} onClose={closeMenu}>
-      {menuItemsWithCloseOnClick}
-    </MuiMenu>
-  );
-};
-
-export const ContextMenu = () => {
-  const anchorPosition = useAnchorPosition();
-  const contextMenuContext = useContextMenuContext();
-
-  if (contextMenuContext === undefined) return <></>;
-
-  const isOpen = Boolean(anchorPosition);
-
-  // create these based on what's set in the context
-  const menuItems = (
-    <MenuItem
-      onClick={() => {
-        deleteNode(contextMenuContext.node.id);
-      }}
-    >
-      Delete node
-    </MenuItem>
-  );
-
-  const menuItemsWithCloseOnClick = addCloseOnClick(closeContextMenu, menuItems);
-
-  return (
-    <MuiMenu
-      anchorReference="anchorPosition"
-      anchorPosition={anchorPosition}
-      open={isOpen}
-      onClose={closeContextMenu}
-    >
       {menuItemsWithCloseOnClick}
     </MuiMenu>
   );
