+++ web/src/pages/index.page.tsx
@@ -3,7 +3,7 @@ import { NextPage } from "next";
 import Head from "next/head";
 import { createContext, useEffect, useState } from "react";
 
-import { ContextMenu } from "../common/components/Menu/Menu";
+import { ContextMenu } from "../common/components/ContextMenu/ContextMenu";
 import { CriteriaTable } from "../modules/topic/components/CriteriaTable/CriteriaTable";
 import { Diagram } from "../modules/topic/components/Diagram/Diagram";
 import { TopicPane } from "../modules/topic/components/Surface/TopicPane";
